Create ca_2021_case_data
This is a .csv file containing the case month, county, age group, current status (I will relabel this column--it really shows if it is a probable case or a laboratory confirmed case),sex, race, and ethnicity. I began with a massive file of all CDC case data for the United States and parsed it down to this. The files I began with are too large to upload.